adjusted class inheritance types

+++ PercentageArc/model/PercentageArcStart.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-public enum PercentageArcStart: CGFloat {
+public enum PercentageArcStart: Int {
     case top = 0
     case right
     case bottom
